dtm: introduce "update identifier".
    - add update type identifier to m0_dtm_update_descr;

    - initialise update operation vector from a new history method:

          m0_dtm_history_ops::hio_update().

    - don't zero the update in m0_dtm_up_init() so that the operation vector can
      be set in adbvance;

+++ dtm/operation.h
@@ -78,7 +78,7 @@ M0_INTERNAL void m0_dtm_reply_unpack(struct m0_dtm_oper *oper,
 				     const struct m0_dtm_oper_descr *reply);
 
 M0_INTERNAL struct m0_dtm_update *m0_dtm_oper_get(const struct m0_dtm_oper *oper,
-						  uint32_t label);
+						  uint8_t label);
 
 /** @} end of dtm group */
